OPP-908 fikser bug med at internflatedecoratoren ikke lenger ble brukt til oppslag. Der ligger logikk for bla validering av fnr/legge bruker i context osv, så den må kalles først.
Oppslag på tråd og oppgave håndteres nå vha vi legger oppgaven i oppgaveResource som automatisk slår opp tråden i meldingerfanen
