+++ src/mock/featureToggle-mock.ts
@@ -11,6 +11,8 @@ export function mockFeatureToggle(toggleId: FeatureToggles): FeatureToggleRespon
             return true;
         case FeatureToggles.SvaksyntModus:
             return false;
+        case FeatureToggles.AvsluttGosysOppgave:
+            return true;
         default:
             return Math.random() > 0.5;
     }
